Refact log components
