+++ site/src/components/Logs/Logs.stories.tsx
@@ -1,0 +1,20 @@
+import type { Meta, StoryObj } from "@storybook/react";
+import { chromatic } from "testHelpers/chromatic";
+import { MockWorkspaceBuildLogs } from "testHelpers/entities";
+import { WorkspaceBuildLogs } from "../../modules/workspaces/WorkspaceBuildLogs/WorkspaceBuildLogs";
+
+const meta: Meta<typeof WorkspaceBuildLogs> = {
+  title: "modules/workspaces/WorkspaceBuildLogs",
+  parameters: { chromatic },
+  component: WorkspaceBuildLogs,
+};
+
+export default meta;
+
+type Story = StoryObj<typeof WorkspaceBuildLogs>;
+
+export const InProgress: Story = {
+  args: {
+    logs: MockWorkspaceBuildLogs.slice(0, 20),
+  },
+};
